% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmicalc.R
\name{cmicalc}
\alias{cmicalc}
\title{Bias-correction factor for Cohen's d}
\usage{
cmicalc(mi)
}
\arguments{
\item{mi}{Cohen's \emph{d}s degrees of freedom (i.e. $n_1+n_2-2$).}
}
\value{
The bias-correction factor.
}
\description{
Bias-correction factor for Cohen's d
}
\examples{
cmicalc(18);
}
