% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_from_2x2.R
\name{d_from_2x2}
\alias{d_from_2x2}
\title{Obtain Cohen's \emph{d} from 2x2 Table Data (Probit Transformation)}
\usage{
d_from_2x2(a, b, c, d, biasCorrect = TRUE)
}
\arguments{
\item{a}{Numerical vector for the number of observations in group 1 with outcome 1.}

\item{b}{Numerical vector for the number of observations in group 1 with outcome 2.}

\item{c}{Numerical vector for the number of observations in group 2 with outcome 1.}

\item{d}{Numerical vector for the number of observations in group 2 with outcome 2.}

\item{biasCorrect}{Logical to indicate if the \emph{d}-values should be bias-corrected. Can also be a vector.}
}
\value{
A numeric vector of Cohen's \code{d} values.
}
\description{
This function computes Cohen's \emph{d} values based on 2x2 table data using the \sQuote{probit transformation}.
}
\examples{
d_from_2x2(a=c(5,12), b=c(35,70), c=c(4,7), d=c(36,70))

}
\references{
Rosenthal, R. (1994). Parametric measures of effect size. In H. Cooper & L. V. Hedges (Eds.), The handbook of research synthesis (1st ed., pp. 231-244). New York: Russell Sage Foundation.

Sanchez-Meca, J., Marin-Martinez, F., & Chacon-Moscoso, S. (2003). Effect-size indices for dichotomized outcomes in meta-analysis. Psychological Methods, 8(4), 448-467.
}
