% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_from_F_oneway.R
\name{d_from_F_oneway}
\alias{d_from_F_oneway}
\title{Obtain Cohen's \emph{d} from one-way F tests}
\usage{
d_from_F_oneway(F1, n1, n2, df2, sign, bias_correct = FALSE)
}
\arguments{
\item{F1}{A numerical vector with one or more \emph{F} values.}

\item{n1, n2}{A numerical vector with the sample sizes of the two groups
formed by the dichotomous variable. Note that the \emph{n}th element of these
vectors must correspond to the \emph{n}th elements of the \code{F1} vector.}

\item{df2}{A numerical vector with one or more values of degrees of freedom
2.}

\item{sign}{Numerical vector to indicate the sign of the d-values
(\code{+1} or \code{-1}).}

\item{bias_correct}{Logical to indicate if the \emph{d}-values should be
bias-corrected. Can also be a vector.}
}
\value{
A data frame with in the first column, Cohen's \code{d} values, and
in the second column, the corresponding variances.
}
\description{
This function converts F statistics from one-way tests to Cohen's \emph{d}.
}
\details{
The formula that is used is the following (see e.g. Thalheimer & Cook,
2002):

\deqn{d= \sqrt{F (\frac{n_1 + n_2}{n_1 n_2}) 
(\frac{n_1 + n_2}{n_1 + n_2 - 2})}}
}
\examples{
escalc::d_from_F_oneway(F1 = 2.828427,
                        n1 = 126,
                        n2 = 89,
                        sign = +1);
escalc::d_from_F_oneway

}
\references{
Thalheimer, W., & Cook, S. (2002, August). \emph{How to calculate
effect sizes from published research articles: A simplified methodology.}

Borenstein, M., Hedges, L. V., Higgins, J. P. T., & Rothstein,
H. R. (2009). \emph{Introduction to Meta-Analysis}. Chichester, UK: John Wiley
& Sons.
}
