% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_from_means.R
\name{d_from_means}
\alias{d_from_means}
\title{Obtain Cohen's \emph{d} from means, standard deviations, and ns}
\usage{
d_from_means(m1, m2, sd1, sd2, n1, n2, bias_correct = FALSE)
}
\arguments{
\item{m1, m2}{A numerical vector with the means of the two groups formed
by the dichotomous variable.}

\item{sd1, sd2}{A numerical vector with the standard deviations of the two
groups formed by the dichotomous variable. Note that the \emph{n}th element of
these vectors must correspond to the \emph{n}th elements of the \code{m1}, \code{m2}
vectors.}

\item{n1, n2}{A numerical vector with the sample sizes of the two groups
formed by the dichotomous variable. Note that the \emph{n}th element of these
vectors must correspond to the \emph{n}th elements of the \code{m1}, \code{m2} vectors.}

\item{bias_correct}{Logical to indicate if the \emph{d}-values should be
bias-corrected. Can also be a vector.}
}
\value{
A data frame with in the first column, Cohen's \code{d} values, and
in the second column, the corresponding variances.
}
\description{
This function converts means, standard deviations, and sample sizes
to Cohen's \emph{d}.
}
\details{
The formula that is used is the following (see e.g. Lakens, 2013):

\deqn{d= \frac{\bar{X}_1 - \bar{X}_1}
{\sqrt{\frac{(n_1 - 1)SD_1^2 + (n_2 - 1)SD_2^2}{n_1 + n_2 - 2}}}}
}
\examples{
escalc::d_from_means(m1 = 2.828427,
                     m2 = 2.123041,
                     sd1 = 0.230101,
                     sd2 = 0.259281,
                     n1 = 126,
                     n2 = 89);

}
\references{
Lakens, D. (2013) Calculating and reporting effect sizes to
facilitate cumulative science: a practical primer for t-tests and ANOVAs.
\emph{Frontiers in Psychology, 4}, p. 863. \doi{10.3389/fpsyg.2013.00863}
}
