% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_from_p_t_in.R
\name{d_from_p_t_in}
\alias{d_from_p_t_in}
\title{Obtain Cohen's \emph{d} from p-value of an independent samples t-test}
\usage{
d_from_p_t_in(p, n1, n2, side, sign, assumeHomoscedacity = TRUE,
  biasCorrect = TRUE)
}
\arguments{
\item{p}{Numerical vector with the (one- or two-sided) p-values.}

\item{n1, n2}{Numerical vectors with the sample sizes of the two groups.}

\item{side}{Numerical vector to indicate the sidedness of the p-values. (\code{1} for a one-sided and \code{2} for a two-sided p-value)}

\item{sign}{Numerical vector to indicate the sign of the d-values (\code{+1} or \code{-1}; only relevant for two-sided p-values; for one-sided p-values, this can be set to \code{NA}).}

\item{assumeHomoscedacity}{Logical to indicate if the p-values come from independent samples t-tests that have assumed homoscedacity (equal variances in the two groups) or not. Can also be a vector.}

\item{biasCorrect}{Logical to indicate if the \emph{d}-values should be bias-corrected. Can also be a vector.}
}
\value{
A numeric vector of Cohen's \code{d} values.
}
\description{
This function converts p-values from independent samples t-tests to Cohen's \emph{d} values.
}
\examples{
d_from_p_t_in(p = .038, n1=22, n2=25, side=2, sign=-1)

}
