% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_from_r_pbs.R
\name{d_from_r_pb}
\alias{d_from_r_pb}
\title{Obtain Cohen's \emph{d} from the point biserial correlation \emph{r}}
\usage{
d_from_r_pb(r_pb, n1, n2, baseRateSensitive = FALSE,
  biasCorrect = FALSE, crosssectionalSampling = FALSE)
}
\arguments{
\item{r_pb}{A numerical vector with one or more point biserial \emph{r} values.}

\item{n1, n2}{A numerical vector with the sample sizes of the two groups
formed by the dichotomous variable. Note that the \emph{n}th element of these
vectors must correspond to the \emph{n}th element of the \code{r_pbs} vector.}

\item{baseRateSensitive}{Whether to compute tyhe base rate sensitive
Cohen's d or not (see McGrath & Meyer, 2006).}

\item{biasCorrect}{Logical to indicate if the \emph{d}-values should be
bias-corrected. Can also be a vector.}
}
\value{
A data frame with in the first column, Cohen's \code{d} values, and
in the second column, the corresponding variances.
}
\description{
This function converts the point biserial correlation \emph{r} to Cohen's \emph{d}.
}
\details{
The formula that is used is the following (see e.g. Borenstein et al., 2009):

\deqn{d= \frac{2 r_{pbs}}{\sqrt{1 - r^2}} }
}
\examples{
escalc::d_from_r_pb(r_pb = .3,
                    n = 100,
                    proportion=.5);

}
\references{
Borenstein, M., Hedges, L. V., Higgins, J. P. T., &
Rothstein, H. R. (2009) \emph{Introduction to Meta-Analysis}, Chichester,
UK: John Wiley & Sons, Ltd.

Also see:

McGrath, R. E. & Meyer, G. J. (2006) When Effect Sizes Disagree:
The Case of \emph{r} and \emph{d}. \emph{Psychological Methods, 11}, 386-401,
doi:\doi{10.1037/1082-989X.11.4.386386}
}
