% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_from_r_pbs.R
\name{d_from_r_pbs}
\alias{d_from_r_pbs}
\title{Obtain Cohen's \emph{d} from the point biserial correlation \emph{r}}
\usage{
d_from_r_pbs(r_pbs, df = NULL, n = NULL, proportion = NULL,
  n1 = NULL, n2 = NULL, baseRateSensitive = FALSE,
  biasCorrected = FALSE)
}
\arguments{
\item{r_pbs}{A numerical vector with one or more point biserial \emph{r} values.}

\item{df, n}{A numerical vector with the degrees of freedom (\code{df}) of \code{r_pbs} or
the total sample size (\code{n}), which is \eqn{df + 2}. Either provide exactly
one of \code{df} or \code{n}, and corresponding \code{proportion}s; \emph{or} provide \code{n1}
and \code{n1}. Note that the \emph{n}th element of the \code{df} and \code{n} vectors must
correspond to the \emph{n}th element of the \code{r_pbs} vector.}

\item{proportion}{A numerical vector with the proportion of participants
in the first (or therefore, implicitly, second) group; must be specified
if \code{df} or \code{n} is specified.  Note that the \emph{n}th element of this vector
must correspond to the \emph{n}th element of the \code{r_pbs} vector.}

\item{n1, n2}{A numerical vector with the sample sizes of the two groups
formed by the dichotomous variable. Note that the \emph{n}th element of these
vectors must correspond to the \emph{n}th element of the \code{r_pbs} vector.}

\item{baseRateSensitive}{Whether to compute}

\item{biasCorrected}{Whether to ------ Wolfgang, Simon, just to check---
are we talking whether to \emph{deliver} d or Hedges' g?}
}
\value{
A data frame with in the first column, Cohen's \code{d} values, and
in the second column, the corresponding variances.
}
\description{
This function converts the point biserial correlation \emph{r} to Cohen's \emph{d}.
}
\details{
The formula that is used is the following (see e.g. Borenstein et al., 2009):

\deqn{d= \frac{2 r_{pbs}}{\sqrt{1 - r^2}} }
}
\examples{
escalc::d_from_r_pbs(r = .3,
                     n = 100,
                     proportion=.5);

}
\references{
Borenstein, M., Hedges, L. V., Higgins, J. P. T., &
Rothstein, H. R. (2009) \emph{Introduction to Meta-Analysis}, Chichester,
UK: John Wiley & Sons, Ltd.

Also see:

McGrath, R. E. & Meyer, G. J. (2006) When Effect Sizes Disagree:
The Case of \emph{r} and \emph{d}. \emph{Psychological Methods, 11}, 386-401,
doi:\doi{10.1037/1082-989X.11.4.386386}
}
