% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_from_t_in.R
\name{d_from_t_in}
\alias{d_from_t_in}
\title{Obtain Cohen's \emph{d} from Student's \emph{t}}
\usage{
d_from_t_in(t, n1, n2, assumeHomoscedacity = TRUE, biasCorrect = FALSE)
}
\arguments{
\item{t}{A numerical vector with one or more \emph{t} values.}

\item{n1, n2}{A numerical vector with the sample sizes of the two groups
formed by the dichotomous variable. Note that the \emph{n}th element of these
vectors must correspond to the \emph{n}th element of the \code{t} vector.}

\item{assumeHomoscedacity}{Whether Student's t is used (assuming equal
variances, or homoscedacity), or Welch's t (assuming unequal variances,
or heteroscedacity). Note that if the variance in the two groups is not
equal, as yet, no method exists for this conversion.}

\item{biasCorrect}{Logical to indicate if the \emph{d}-values should be
bias-corrected. Can also be a vector.}
}
\value{
A data frame with in the first column, Cohen's \code{d} values, and
in the second column, the corresponding variances.
}
\description{
This function converts Student's \emph{t} to Cohen's \emph{d}.
}
\details{
The formula that is used is the following (see e.g. Lakens, 2013):

\deqn{d= t \sqrt{(\frac{1}{n_1} + \frac{1}{n_2})}}{d=t*sqrt(1/n1 + 1/n2))}
}
\examples{
escalc::d_from_t_in(t = 2.828427,
                    n1 = 126,
                    n2 = 89);

}
\references{
Lakens, D. (2013) Calculating and reporting effect sizes to
facilitate cumulative science: a practical primer for t-tests and ANOVAs.
\emph{Frontiers in Psychology, 4}, p. 863. \doi{10.3389/fpsyg.2013.00863}
}
