% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_to_d.R
\name{t_to_d}
\alias{t_to_d}
\title{Student's \emph{t} to Cohen's \emph{d}}
\usage{
t_to_d(t, df, n1, n2, proportion = 0.5)
}
\arguments{
\item{t}{The \emph{t} value.}

\item{df}{The degrees of freedom of \code{t}; either provide this \code{proportion},
or \code{n1} and \code{n1}.}

\item{n1, n2}{The sample sizes of the two groups formed by the dichotomous
variable.}

\item{proportion}{The proportion of participants in the first (or second)
group; must be specified if \code{df} is specified.}
}
\value{
A numeric vector of Cohen's \code{d} values.
}
\description{
This function converts Student's \emph{t} to Cohen's \emph{d}.
}
\details{
The formula that is used is the following (see e.g. Lakens, 2013):

\deqn{d= t \sqrt{(\frac{1}{n_1} + \frac{1}{n_2})}}{d=t*sqrt(1/n1 + 1/n2))}
}
\examples{
escalc::t_to_d(t = 2.85, df = 128);

}
\references{
Lakens, D. (2013) Calculating and reporting effect sizes to
facilitate cumulative science: a practical primer for t-tests and ANOVAs.
\emph{Frontiers in Psychology, 4}, p. 863. \doi{10.3389/fpsyg.2013.00863}
}
